+++ storage/backup/images_backup_20260829_005546.sql
@@ -1,0 +1,84 @@
+mysqldump: [Warning] Using a password on the command line interface can be insecure.
+-- MySQL dump 10.13  Distrib 5.7.44, for Linux (x86_64)
+--
+-- Host: localhost    Database: pic_5276_net
+-- ------------------------------------------------------
+-- Server version	5.7.44-log
+
+/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
+/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
+/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
+/*!40101 SET NAMES utf8 */;
+/*!40103 SET @OLD_TIME_ZONE=@@TIME_ZONE */;
+/*!40103 SET TIME_ZONE='+00:00' */;
+/*!40014 SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0 */;
+/*!40014 SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0 */;
+/*!40101 SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='NO_AUTO_VALUE_ON_ZERO' */;
+/*!40111 SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0 */;
+mysqldump: Error: 'Access denied; you need (at least one of) the PROCESS privilege(s) for this operation' when trying to dump tablespaces
+
+--
+-- Table structure for table `images`
+--
+
+DROP TABLE IF EXISTS `images`;
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!40101 SET character_set_client = utf8 */;
+CREATE TABLE `images` (
+  `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
+  `uuid` varchar(36) NOT NULL,
+  `user_id` int(10) unsigned NOT NULL,
+  `storage_id` int(10) unsigned NOT NULL,
+  `folder_id` int(10) unsigned DEFAULT NULL,
+  `original_name` varchar(255) NOT NULL,
+  `stored_name` varchar(255) NOT NULL,
+  `extension` varchar(16) NOT NULL,
+  `mime_type` varchar(64) NOT NULL,
+  `width` int(10) unsigned DEFAULT NULL,
+  `height` int(10) unsigned DEFAULT NULL,
+  `original_size` bigint(20) unsigned NOT NULL DEFAULT '0',
+  `final_size` bigint(20) unsigned NOT NULL DEFAULT '0',
+  `compression_ratio` decimal(5,2) DEFAULT '0.00',
+  `sha256` varchar(64) NOT NULL,
+  `storage_path` varchar(512) NOT NULL,
+  `public_url` varchar(512) NOT NULL,
+  `status` enum('active','expired','recycle') NOT NULL DEFAULT 'active',
+  `is_public` tinyint(1) NOT NULL DEFAULT '1',
+  `tags` varchar(500) DEFAULT NULL,
+  `expires_at` datetime DEFAULT NULL,
+  `created_at` datetime NOT NULL,
+  `updated_at` datetime DEFAULT NULL,
+  `deleted_at` datetime DEFAULT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `uk_uuid` (`uuid`),
+  UNIQUE KEY `uk_sha256_user` (`sha256`,`user_id`),
+  KEY `idx_user_id` (`user_id`),
+  KEY `idx_folder_id` (`folder_id`),
+  KEY `idx_status` (`status`),
+  KEY `idx_expires_at` (`expires_at`),
+  KEY `idx_deleted_at` (`deleted_at`),
+  KEY `idx_created_at` (`created_at`)
+) ENGINE=InnoDB AUTO_INCREMENT=36 DEFAULT CHARSET=utf8mb4;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Dumping data for table `images`
+--
+-- WHERE:  id <= 25
+
+LOCK TABLES `images` WRITE;
+/*!40000 ALTER TABLE `images` DISABLE KEYS */;
+INSERT INTO `images` VALUES (24,'639beed3-8628-4edc-a655-e9edc7198d44',3,2,NULL,'2024-12-31_193350.jpg','JH1Nn9k.jpg','jpg','image/jpeg',1087,1600,112611,110836,0.98,'655a84213270a307de6ba1640b021ead80965f81e1450dc84ee612089ff3d87f','JH1Nn9k.jpg','https://pic.5276.net/img/JH1Nn9k.jpg','active',1,NULL,NULL,'2026-08-28 23:21:06','2026-08-28 23:21:06',NULL);
+/*!40000 ALTER TABLE `images` ENABLE KEYS */;
+UNLOCK TABLES;
+/*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
+
+/*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
+/*!40014 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS */;
+/*!40014 SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS */;
+/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
+/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
+/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
+/*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
+
+-- Dump completed on 2026-08-29  0:55:46